table

+++ 0x02-functions_nested_loops/main.h
@@ -6,7 +6,7 @@
 void print_to_98(int n);
 void times_table(void);
 void jack_bauer(void);
-int print_last_digit(int x);
+int print_last_digit(long int x);
 int _abs(int x);
 int print_sign(int n);
 int _isalpha(int c);
